Move shared plugins to `shared-react`
- git ignore NX cache
- also fix NoteNode onClick
