+++ .eslintignore
@@ -45,3 +45,6 @@ testem.log
 # System Files
 .DS_Store
 Thumbs.db
+
+# NX
+.nx/cache
