+++ .gitignore
@@ -40,3 +40,6 @@ testem.log
 # System Files
 .DS_Store
 Thumbs.db
+
+# NX
+.nx/cache
