+++ packages/shared-react/adaptors/editor-adaptor.model.ts
