+++ packages/shared-react/index.js
@@ -1,0 +1,1 @@
+export default null;
