+++ packages/shared-react/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "shared-react",
+  "private": "true",
+  "keywords": [
+    "react",
+    "lexical",
+    "editor",
+    "rich-text"
+  ],
+  "scripts": {},
+  "license": "MIT",
+  "version": "0.0.1",
+  "dependencies": {
+    "@lexical/react": "^0.12.2",
+    "@lexical/utils": "^0.12.2",
+    "lexical": "0.12.2",
+    "shared": "*"
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.35",
+    "@types/react-dom": "^18.2.14",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "peerDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/abelpz/bible-lexical-editor",
+    "directory": "packages/shared-react"
+  }
+}
