+++ packages/shared-react/project.json
@@ -1,0 +1,20 @@
+{
+  "name": "shared-react",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "{projectRoot}",
+  "projectType": "package",
+  "targets": {
+    "build": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "tsc -p tsconfig.json",
+        "cwd": "{projectRoot}",
+        "forwardAllArgs": false
+      }
+    },
+    "typecheck": {},
+    "test": {},
+    "lint": {}
+  },
+  "tags": []
+}
